% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_safe_environment.R
\name{create_safe_environment}
\alias{create_safe_environment}
\title{Create a safe environment}
\usage{
create_safe_environment(dplyr = FALSE, plot = FALSE)
}
\arguments{
\item{dplyr}{Load a collections of \code{dplyr} functions in the environment.}

\item{plot}{Load a collection of \code{graphics} and \code{ggplot2} functions in the
environment.}
}
\value{
An environment that uses the empty environment (\code{emptyenv()}) as
its parent. The environment serves as a "whitelist" to prevent users
of \link{evaluator}\code{$eval()} to invoke malicious operations.
}
\description{
Creates an environment where the safe expressions are to be evaluated.
}
\examples{
safe_env <- create_safe_environment()
eval(parse(text = "2 + 2"), safe_env)

\dontrun{
## trigger an error
eval(parse(text = "system('mkdir(test)')"), safe_env)
}

}
