% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\docType{data}
\name{evaluator}
\alias{evaluator}
\title{evaluate strings}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
evaluator
}
\description{
evaluate strings
}
\keyword{datasets}
